extend archiver_args->{u,g}id for 16-bit systems

+++ archive.h
@@ -15,8 +15,8 @@ struct archiver_args {
 	unsigned int worktree_attributes : 1;
 	unsigned int convert : 1;
 	int compression_level;
-	int uid;
-	int gid;
+	unsigned long uid;
+	unsigned long gid;
 	const char *uname;
 	const char *gname;
 };
